feat : 예약 조회, 에러처리 로직 변경
