+++ src/main/java/com/ssafy/enjoytrip/global/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.ssafy.enjoytrip.global;
+package com.ssafy.enjoytrip.global.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
