+++ src/main/java/com/ssafy/enjoytrip/member/model/service/MemberServiceImpl.java
@@ -1,6 +1,6 @@
 package com.ssafy.enjoytrip.member.model.service;
 
-import com.ssafy.enjoytrip.global.ErrorCode;
+import com.ssafy.enjoytrip.global.exception.ErrorCode;
 import com.ssafy.enjoytrip.global.service.EmailService;
 import com.ssafy.enjoytrip.member.exception.MemberException;
 import com.ssafy.enjoytrip.member.model.dto.MemberJoinDto;
