+++ src/main/java/com/ssafy/enjoytrip/room/exception/RoomException.java
@@ -1,6 +1,6 @@
 package com.ssafy.enjoytrip.room.exception;
 
-import com.ssafy.enjoytrip.global.ErrorCode;
+import com.ssafy.enjoytrip.global.exception.ErrorCode;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
