+++ src/main/java/com/ssafy/enjoytrip/transaction/Exception/TransactionException.java
@@ -1,4 +1,4 @@
-package com.ssafy.enjoytrip.reservation.exception;
+package com.ssafy.enjoytrip.transaction.Exception;
 
 import com.ssafy.enjoytrip.global.exception.ErrorCode;
 import lombok.AllArgsConstructor;
@@ -6,8 +6,9 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class ReservationException extends RuntimeException{
+public class TransactionException extends RuntimeException{
 
     private ErrorCode errorCode;
     private String message;
+
 }
