+++ src/main/resources/mapper/transaction.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.ssafy.enjoytrip.transaction.model.mapper.TransactionMapper">
+
+  <resultMap type="transaction" id="transaction">
+    <result column="transaction_id" property="id"/>
+    <result column="member_id" property="memberId"/>
+    <result column="transaction_amount" property="amount"/>
+    <result column="description" property="description"/>
+    <result column="createdat" property="createdat"/>
+    <result column="updatedat" property="updatedat"/>
+  </resultMap>
+
+  <select id="findAllByMemberId" resultMap="transaction">
+    SELECT *
+    FROM transaction;
+  </select>
+
+</mapper>
